requirements.txt
pdfplumber==0.10.2
svgwrite==1.4.3
pdf2image==1.16.3
Pillow==10.0.0
numpy==1.24.3
img2pdf==0.4.4
transformers
jiwer
pytesseract==0.3.10
accelerate>=0.26.0

# To be able to use apple silicon gpu
torch>=2.0.0  # Version 2.0.0+ has good MPS support
torchvision>=0.15.0
torchaudio>=2.0.0


# 
kraken
cmake

# for trocr
opencv-python


pandas>=1.3.0
matplotlib>=3.5.0
seaborn>=0.11.0
tqdm>=4.62.0

# Synthetic data generator
trdg